Начал Комментарий Реализовал чтоб мог выбирать столик через картинку Добавил Алгоритм Ант, Добавил роли теперь нужно добить эту часть , закончил перевод нужно роли добавить, закончил админку, закончил Телеграм, теперь Телеграм начать и исправлять весь проект , Реализован PersonDisplay. потом добить телеграм Остался один метод в Admin закончив преступать к стилю, тесту и исправления Залили в домен, теперь нужно добить админ Нужно с Борисом поговорить и залить в Домен, добавил второй метод для админки,Успешно собрал варку, Релизовал MultiЯзычность и создал метод для изменения языка, теперь нужно словарь заполнять, сделал admin server и создал один метод еще есть 4 метода , нужно начать админку, после делать исправления по всему проекту

+++ diploma.controller/src/kz/greetgo/diploma/controller/register/model/Orders.java
@@ -17,7 +17,7 @@ public class Orders {
 
 	public String pMethod;
 
-	public Integer gTotal;
+	public Long gTotal;
 
 	public Integer bookingId;
